+++ database/migrations/2021_07_20_093222_create_orders_table.php
@@ -16,7 +16,7 @@ class CreateOrdersTable extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->string('number');
-            $table->float('amount');
+            $table->decimal('amount');
             $table->foreignId('user_id')->constrained;
             $table->timestamps();
         });
